modified:   __pycache__/brute_force.cpython-311.pyc 	modified:   __pycache__/file_manager.cpython-311.pyc 	new file:   __pycache__/heap_manager.cpython-311.pyc 	modified:   __pycache__/round_robin.cpython-311.pyc 	modified:   brute_force.py 	modified:   file_manager.py 	modified:   main.py 	modified:   round_robin.py

+++ file_manager.py
@@ -1,4 +1,7 @@
 
+from decimal import Decimal
+
+
 class file_manager:
     def __init__(self,path):
         self.path = path
@@ -11,5 +14,5 @@ class file_manager:
 
         for i in range(len(lines)):
             split_line = lines[i].split()
-            R.append(float(split_line[1]))
+            R.append(Decimal(split_line[1]))
         return R
